requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.4
django-autoslug==1.9.8
django-cities==0.6.1
django-cities-light==3.9.2
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.13.0
django-humanize==0.1.2
django-model-utils==4.3.1
django-widget-tweaks==1.4.12
GDAL @ file:///C:/expatproj/GDAL-3.3.3-cp39-cp39-win_amd64.whl
humanize==4.6.0
idna==3.4
Pillow==9.3.0
progressbar2==4.2.0
psycopg2==2.9.5
python-decouple==3.5
python-utils==3.4.5
pytz==2022.7
razorpay==1.3.0
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
swapper==1.3.0
tqdm==4.64.1
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.13
